Move Package definition to a Base module
